README - adding to the master branch
How to commit to master branch so we can work collaboratively.

Click on the project.c file under "<> code".
Click on the pencil to the right and edit the file.

After making your changes, go to the bottom and add a description of the change and commit to the master branch. 
The next page will ask if you want to merge pull request (or something like that).
This will send a request to the people with write permission or automatically commit to it. 
Check the project file to make sure it was committed.